fix(windows): enforce owned filesystem boundaries
